refactor(photo): Reorganize photo data access layer
Moves the data source logic into a more structured `provider` directory.

- Renames `query.ts` to `database.ts` to better reflect its role of direct database interaction.
- Establishes a clear separation between the data access interface and its implementation, preparing for the addition of new data sources like Immich.
